Implement persistent sessions.
Persistent sessions will allow a newly started cloud-connector to
get notified of client connections and disconnections which happened
while the cloud-connector was down.

There appears to be a bit of a race condition in the paho client.
The paho client appears to receive messages from the broker
before the subscriber is completely "wired up".  Added a default
publish handler to work around that issue.

The issue and workaround are mentioned in the Common Problems section here:
https://pkg.go.dev/github.com/eclipse/paho.mqtt.golang