[api] Add disconnect endpoint to API spec
Gathered what I could from the code. Looks like we just send an OK
status and empty JSON if the disconnect is successful. I don't see items
in the spec that account for the 400s, so I didn't put that in.

Let me know if there are additional needs or if I misread anything

RHCLOUD-12532

Signed-off-by: Stephen Adams <tsadams@gmail.com>